add changing user functionality Currently not working properly because I don't know how to write changes to file, so I can read from it later...
