+++ project1/UserClass.py
@@ -5,3 +5,4 @@ class User:
         self.email = email
         self.phone = phone
         self.uuid = uuid
+
